feat: improved NoteControllers

+++ src/main/kotlin/ru/normno/backendcourse/spring_boot_course/controllers/NoteControllers.kt
@@ -1,6 +1,7 @@
 package ru.normno.backendcourse.spring_boot_course.controllers
 
 import org.bson.types.ObjectId
+import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.web.bind.annotation.*
 import ru.normno.backendcourse.spring_boot_course.database.model.Note
 import ru.normno.backendcourse.spring_boot_course.database.repository.NoteRepository
@@ -31,6 +32,7 @@ class NoteControllers(
     fun save(
         @RequestBody body: NoteRequest
     ): NoteResponse {
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
         val note = repository.save(
             Note(
                 id = body.id?.let { ObjectId(it) } ?: ObjectId.get(),
@@ -38,7 +40,7 @@ class NoteControllers(
                 content = body.content,
                 color = body.color,
                 createdAt = Instant.now(),
-                ownerId = ObjectId(),
+                ownerId = ObjectId(ownerId),
             )
         )
 
@@ -52,9 +54,8 @@ class NoteControllers(
     }
 
     @GetMapping
-    fun findByOwnerId(
-        @RequestParam(required = true) ownerId: String,
-    ): List<NoteResponse> {
+    fun findByOwnerId(): List<NoteResponse> {
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
         return repository.findByOwnerId(ObjectId(ownerId)).map {
             it.toResponse()
         }
@@ -74,6 +75,11 @@ class NoteControllers(
     fun deleteById(
         @PathVariable id: String,
     ) {
-        repository.deleteById(ObjectId(id))
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
+        val note = repository.findById(ObjectId(id)).orElseThrow {
+            IllegalArgumentException("Note not found")
+        }
+        if (note.ownerId.toHexString() == ownerId)
+            repository.deleteById(ObjectId(id))
     }
 }
